Resolve conflicts against the restructured deep_sleep tests: dev now uses
test_deep_sleep3.yaml for its ext1 on_wake test, so the nested
wakeup_pin_mode config moves to test_deep_sleep4.yaml.

+++ esphome/components/sound_level/sound_level.cpp
@@ -121,7 +121,7 @@ void SoundLevelComponent::loop() {
   if (this->sample_count_ == samples_in_window) {
     // Processed enough samples for the measurement window, compute and publish the sensor values
     if (this->peak_sensor_ != nullptr) {
-      const float peak_db = 10.0f * log10(static_cast<float>(this->squared_peak_) / MAX_SAMPLE_SQUARED_DENOMINATOR);
+      const float peak_db = 10.0f * log10f(static_cast<float>(this->squared_peak_) / MAX_SAMPLE_SQUARED_DENOMINATOR);
       this->peak_sensor_->publish_state(peak_db);
 
       this->squared_peak_ = 0;  // reset accumulator
